[UsdNpr] hydra 2.0 contour prim work in progress

+++ plugins/usdNprImaging/mesh.cpp
@@ -157,7 +157,7 @@ void UsdNprHalfEdgeMesh::Init(const UsdGeomMesh& mesh, const UsdTimeCode& timeCo
 
   UsdNprHalfEdge* halfEdge = &_halfEdges[0];
   const int* indices = &faceVertexIndices[0];
-  size_t halfEdgeIndex = 0;
+  uint32_t halfEdgeIndex = 0;
 
   // for each face build half-edges and insert in map
   for(int faceIndex = 0; faceIndex < _numPolygons; ++ faceIndex) {
